+++ Components/Settings/Settings.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { View, Text, SafeAreaView } from 'react-native'
+const Settings = () => {
+  return (
+    <>
+      <SafeAreaView>
+        <View>
+          <Text>Settings</Text>
+        </View>
+      </SafeAreaView>
+      
+    </>
+    
+  )
+}
+
+export default Settings
