+++ helpers/mockData.js
@@ -2,17 +2,20 @@ const mockItemData = [
   {
     title: 'Pepperoni',
     isNew: false,
+    price: '350$',
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Mushroom Delight',
+    price: '350$',
     isNew: true,
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Sausage Feast',
+    price: '350$',
     isNew: false,
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
@@ -20,17 +23,20 @@ const mockItemData = [
   {
     title: 'Veggie Supreme',
     isNew: true,
+    price: '350$',
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Cheesy Pizza',
+    price: '350$',
     isNew: false,
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Classic Pepperoni',
+    price: '350$',
     isNew: true,
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
@@ -44,17 +50,20 @@ const mockItemData = [
   {
     title: 'Margherita Magic',
     isNew: true,
+    price: '350$',
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Hawaiian Retreat',
     isNew: true,
+    price: '350$',
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
   },
   {
     title: 'Seafood Symphony',
+    price: '350$',
     isNew: false,
     image: 'https://img.freepik.com/free-photo/top-view-pepperoni-pizza-with-mushroom-sausages-bell-pepper-olive-corn-black-wooden_141793-2158.jpg',
     description: 'Long title Long title Long title Long title Long title Long title Long title Long title ',
