+++ package.json
@@ -9,12 +9,25 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-navigation/bottom-tabs": "^6.5.12",
+    "@react-navigation/bottom-tabs": "^6.5.19",
+    "@react-navigation/drawer": "^6.6.14",
+    "@react-navigation/material-bottom-tabs": "^6.2.27",
+    "@react-navigation/native": "^6.1.16",
+    "@react-navigation/native-stack": "^6.9.25",
+    "@react-navigation/stack": "^6.3.28",
     "expo": "~50.0.7",
     "expo-status-bar": "~1.11.1",
+    "i": "^0.3.7",
+    "npm": "^10.5.0",
     "react": "18.2.0",
     "react-native": "0.73.4",
-    "react-native-btr": "^2.2.1"
+    "react-native-btr": "^2.2.1",
+    "react-native-gesture-handler": "^2.15.0",
+    "react-native-paper": "^5.12.3",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.29.0",
+    "react-native-vector-icons": "^10.0.3",
+    "react-native-reanimated": "~3.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
